requirements.txt
tensorflow-gpu==2.4.0
pandas==1.3.4
matplotlib==3.2.1
